SimClient and SimServer now handle simulation problems and return a .fail file in case of error.

+++ spicelib/client_server/srv_sim_runner.py
@@ -20,21 +20,43 @@
 # -------------------------------------------------------------------------------
 import threading
 import time
-from typing import Union, Any
+from typing import Union, Any, Optional
 from pathlib import Path
 import zipfile
 import logging
-_logger = logging.getLogger("spicelib.ServerSimRunner")
 
 from ..sim.sim_runner import SimRunner
 from ..editor.base_editor import BaseEditor
 
+_logger = logging.getLogger("spicelib.ServerSimRunner")
 
-def zip_files(raw_filename: Path, log_filename: Path):
-    zip_filename = raw_filename.with_suffix('.zip')
+
+def zip_files(raw_filename: Optional[Path], log_filename: Optional[Path]) -> Optional[Path]:
+    """Zips the raw and log files into a single zip file.
+
+    :param raw_filename: The path to the raw file.
+    :type raw_filename: Optional[pathlib.Path]
+    :param log_filename: The path to the log file.
+    :type log_filename: Optional[pathlib.Path]
+    :return: The path to the zip file or None if no files were provided.
+    :rtype: Optional[Path]
+    """
+    zip_filename = None
+    if isinstance(raw_filename, Path) and raw_filename.exists():
+        zip_filename = raw_filename.with_suffix('.zip')
+    else:
+        if isinstance(log_filename, Path) and log_filename.exists():
+            zip_filename = log_filename.with_suffix('.zip')
+            
+    if zip_filename is None:
+        _logger.warning("No files are available for zipping. Returning None.")
+        return None
+
     with zipfile.ZipFile(zip_filename, 'w', zipfile.ZIP_DEFLATED) as zip_file:
-        zip_file.write(raw_filename)
-        zip_file.write(log_filename)
+        if isinstance(raw_filename, Path) and raw_filename.exists():
+            zip_file.write(raw_filename)
+        if isinstance(log_filename, Path) and log_filename.exists():
+            zip_file.write(log_filename)
     return zip_filename
 
 
@@ -49,8 +71,8 @@ class ServerSimRunner(threading.Thread):
     by this class.
     """
 
-    def __init__(self, parallel_sims: int = 4, timeout: float = None, verbose=False,
-                 output_folder: str = None, simulator=None):
+    def __init__(self, parallel_sims: int = 4, timeout: float = 600.0, verbose=False,
+                 output_folder: Optional[str] = None, simulator=None):
         super().__init__(name="SimManager")
         self.runner = SimRunner(simulator=simulator, parallel_sims=parallel_sims, timeout=timeout,
                                 verbose=verbose, output_folder=output_folder)
@@ -92,7 +114,7 @@ class ServerSimRunner(threading.Thread):
         self.runner.wait_completion()
         self.runner.cleanup_files()  # Delete things that have been left behind
 
-    def add_simulation(self, netlist: Union[str, Path, BaseEditor], *, timeout: float = None) -> int:
+    def add_simulation(self, netlist: Union[str, Path, BaseEditor], *, timeout: Optional[float] = None) -> int:
         """
         Adding a simulation to the list of simulations to be run. The function will return the runno of the simulation
         or -1 if the simulation could not be started.
@@ -102,7 +124,7 @@ class ServerSimRunner(threading.Thread):
         :return: The runno of the simulation or -1 if the simulation could not be started
         """
         _logger.debug(f"starting Simulation of {netlist}")
-        task = self.runner.run(netlist, wait_resource=True, timeout=timeout, callback=zip_files)
+        task = self.runner.run(netlist, wait_resource=True, timeout=timeout, callback=zip_files, callback_on_error=True)
         if task is None:
             _logger.error(f"Failed to start task {netlist}")
             return -1
@@ -113,21 +135,28 @@ class ServerSimRunner(threading.Thread):
     def _erase_files_and_info(self, pos):
         task = self.completed_tasks[pos]
         for filename in ('circuit', 'log', 'raw', 'zipfile'):
-            f = task[filename]
-            if f.exists():
-                _logger.info(f"deleting {f}")
-                f.unlink()
+            if filename in task:
+                f = task[filename]
+                if f is not None and f.exists():
+                    _logger.debug(f"deleting file {f}")
+                    try:
+                        f.unlink()
+                        _logger.debug(f"deleted file {f}")
+                    except Exception as e:
+                        _logger.error(f"Error deleting file {f}: {e}")
         del self.completed_tasks[pos]
 
     def erase_files_of_runno(self, runno):
         """Will delete all files related with a completed task. Will also delete information on the completed_tasks
         attribute."""
+        _logger.debug(f"deleting files of run {runno}")
         for i, task_info in enumerate(self.completed_tasks):
             if task_info['runno'] == runno:
                 self._erase_files_and_info(i)
                 break
 
     def cleanup_completed(self):
+        _logger.debug("deleting all files of all completed tasks")
         while len(self.completed_tasks):
             self._erase_files_and_info(0)
